+++ config/settings/base.py
@@ -362,6 +362,7 @@ SPECTACULAR_SETTINGS = {
 # ------------------------------------------------------------------------------
 
 # Knox settings
+KNOX_TOKEN_MODEL = 'tokens.SODARAuthToken'
 TOKEN_TTL = None
 
 # Settings for HTTP AuthBasic