+++ tokens/tests/test_permissions_api.py
@@ -1,3 +1,0 @@
-"""Tests for REST API view permissions in the tokens app"""
-
-# TODO: Add test