+++ tokens/tests/test_ui.py
@@ -6,14 +6,14 @@ from django.test import override_settings
 from django.urls import reverse
 from django.utils import timezone
 
-from knox.models import AuthToken
-
 from selenium.webdriver.common.by import By
 
 # Projectroles dependency
 from projectroles.app_settings import AppSettingAPI
 from projectroles.models import SODAR_CONSTANTS
 from projectroles.tests.test_ui import UITestBase
+
+from tokens.tests.test_models import SODARAuthTokenMixin
 
 
 app_settings = AppSettingAPI()
@@ -26,7 +26,7 @@ PROJECT_TYPE_CATEGORY = SODAR_CONSTANTS['PROJECT_TYPE_CATEGORY']
 APP_NAME_PR = 'projectroles'
 
 
-class TestUserTokenListView(UITestBase):
+class TestUserTokenListView(SODARAuthTokenMixin, UITestBase):
     """Tests for UserTokenListView"""
 
     def setUp(self):
@@ -39,12 +39,10 @@ class TestUserTokenListView(UITestBase):
         self.regular_user = self.make_user('regular_user', False)
 
         # Create tokens
-        self.token = AuthToken.objects.create(
-            self.regular_user, timedelta(days=5)
+        self.token = self.make_token(
+            self.regular_user, expiry=timedelta(days=5)
         )[0]
-        self.token_no_expiry = AuthToken.objects.create(
-            self.regular_user, None
-        )[0]
+        self.token_no_expiry = self.make_token(self.regular_user)[0]
         self.url = reverse('tokens:list')
 
     def test_create_button(self):
@@ -99,7 +97,7 @@ class TestUserTokenListView(UITestBase):
             expected, self.url, 'sodar-tk-list-item', 'class'
         )
         self.assert_element_count(
-            expected, self.url, 'sodar-tk-item-dropdown', 'class'
+            expected, self.url, 'sodar-tk-delete-btn', 'class'
         )
 
     def test_list_items_read_only(self):
@@ -111,7 +109,7 @@ class TestUserTokenListView(UITestBase):
         )
         expected = [(self.superuser, 0), (self.regular_user, 0)]
         self.assert_element_count(
-            expected, self.url, 'sodar-tk-item-dropdown', 'class'
+            expected, self.url, 'sodar-tk-delete-btn', 'class'
         )
 
     def test_list_expiry(self):
@@ -119,9 +117,7 @@ class TestUserTokenListView(UITestBase):
         self.login_and_redirect(self.regular_user, self.url)
         items = self.selenium.find_elements(By.CLASS_NAME, 'sodar-tk-list-item')
         self.assertEqual(len(items), 2)
-        expiry_time = timezone.localtime(self.token.expiry).strftime(
-            '%Y-%m-%d %H:%M'
-        )
+        expiry_time = timezone.localtime(self.token.expiry).strftime('%Y-%m-%d')
         values = []
         for item in items:
             values.append(item.find_elements(By.TAG_NAME, 'td')[2].text)